rename graph.cpp to odgi.cpp
