+++ CMakeLists.txt
@@ -224,7 +224,7 @@ set(CMAKE_BUILD_TYPE Release)
 
 # set up our target executable and specify its dependencies and includes
 add_executable(odgi
-  ${CMAKE_SOURCE_DIR}/src/graph.cpp
+  ${CMAKE_SOURCE_DIR}/src/odgi.cpp
   ${CMAKE_SOURCE_DIR}/src/main.cpp
   ${CMAKE_SOURCE_DIR}/src/crash.cpp
   ${CMAKE_SOURCE_DIR}/src/position.cpp
@@ -311,10 +311,7 @@ target_link_libraries(odgi
   "${sonLib_LIB}/3EdgeConnected.a"
   "${sonLib_LIB}/sonLib.a"
   "-ldl"
-#  "-ltcmalloc_minimal"
   )
-
-#set (CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -ltcmalloc")
 
 if (APPLE)
 elseif (TRUE)
