+++ src/unittest/simplify.cpp
@@ -3,7 +3,7 @@
 #include <handlegraph/handle_graph.hpp>
 #include <handlegraph/handle_graph.hpp>
 #include <handlegraph/util.hpp>
-#include "graph.hpp"
+#include "odgi.hpp"
 #include "algorithms/simple_components.hpp"
 #include "algorithms/topological_sort.hpp"
